[Script #396] framer motion 설치
