+++ utils/socket/socket.ts
@@ -2,6 +2,6 @@
 
 import { Socket, io } from "socket.io-client";
 
-export const socket: Socket = io(process.env.NEXT_PUBLIC_SOCKET_SERVER_URL!, {
+export const socket: Socket = io(process.env.NEXT_PUBLIC_SOCKET_DEV_SERVER_URL!, {
   autoConnect: false
 });
